+++ tests/TestModel.php
@@ -8,6 +8,4 @@ class TestModel extends Model
 {
 	protected $table = 'test_models';
 
-	
-
 }