+++ tests/TestSlugFieldModel.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Extended\Tests;
+
+use Extended\Traits\HasSlug;
+use Illuminate\Database\Eloquent\Model;
+
+class TestSlugFieldModel extends Model
+{
+	use HasSlug;
+
+	protected $table = 'test_slug_models';
+
+	protected $findSlugField = 'uri';
+	protected $slugField = 'uri';
+
+}