+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-eloquent-extended` will be documented in this file
 
+## 1.0.2 - 2020-04-23
+
+- add the HasSlug trait
+- remove Meta as it was just an extension of the Content fields
+
 ## 1.0.1 - 2020-04-16
 
 - renamed to eloquent-extended